Class and namespace renames
