+++ WebApplication/WorkoutTracker.Api/Controllers/Concrete/WorkoutTemplateController.cs
@@ -1,26 +1,27 @@
 ﻿using System.Collections.Generic;
 using System.Web.Http;
 using WorkoutTracker.Api.Controllers.Abstract;
+using WorkoutTracker.Core.Implementation.ActionDispatchers.Abstract;
 using WorkoutTracker.Core.Implementation.Actions.WorkoutTemplateActions;
-using WorkoutTracker.Core.Implementation.CommandDispatchers.Abstract;
 using WorkoutTracker.Core.Implementation.Domain;
 using WorkoutTracker.Core.Implementation.Queries;
+using WorkoutTracker.Core.Implementation.Queries.Concrete;
 using WorkoutTracker.Core.Implementation.QueryHandlers.Abstract;
 
 namespace WorkoutTracker.Api.Controllers.Concrete
 {
     public class WorkoutTemplateController : BaseController
     {
-        private readonly IWorkoutTemplateCommandDispatcher _commandDispatcher;
+        private readonly IWorkoutTemplateActionDispatcher _actionDispatcher;
         private readonly IQueryHandler<WorkoutTemplateQuery, IEnumerable<WorkoutTemplate>> _queryHandler;
 
         public WorkoutTemplateController() { }
 
         public WorkoutTemplateController(
-            IWorkoutTemplateCommandDispatcher commandDispatcher,
+            IWorkoutTemplateActionDispatcher actionDispatcher,
             IQueryHandler<WorkoutTemplateQuery, IEnumerable<WorkoutTemplate>> queryHandler)
         {
-            _commandDispatcher = commandDispatcher;
+            _actionDispatcher = actionDispatcher;
             _queryHandler = queryHandler;
         }
 
@@ -34,7 +35,7 @@ namespace WorkoutTracker.Api.Controllers.Concrete
 
         public IHttpActionResult Post(AddWorkoutTemplateAction action)
         {
-            return Result(_commandDispatcher.Dispatch(action));
+            return Result(_actionDispatcher.Dispatch(action));
         }
     }
 }
