+++ WebApplication/WorkoutTracker.Contracts/ActionHandlers/Concrete/WorkoutTemplateActionHandlers/AddWorkoutTemplateActionHandler.cs
@@ -1,10 +1,9 @@
-﻿using System.Linq;
-using WorkoutTracker.Core.Implementation.ActionHandlers.Abstract;
+﻿using WorkoutTracker.Core.Implementation.ActionHandlers.Abstract;
 using WorkoutTracker.Core.Implementation.Actions.WorkoutTemplateActions;
 using WorkoutTracker.Core.Implementation.DbContexts.Abstract;
 using WorkoutTracker.Core.Implementation.Domain;
 
-namespace WorkoutTracker.Core.Implementation.ActionHandlers.Concrete.WorkoutTemplateCommands
+namespace WorkoutTracker.Core.Implementation.ActionHandlers.Concrete.WorkoutTemplateActionHandlers
 {
     public class AddWorkoutTemplateActionHandler : IActionHandler<AddWorkoutTemplateAction>
     {
