+++ WebApplication/WorkoutTracker.Contracts/Queries/Concrete/WorkoutTemplateQuery.cs
@@ -2,7 +2,7 @@
 using WorkoutTracker.Core.Implementation.Domain;
 using WorkoutTracker.Core.Implementation.Queries.Abstract;
 
-namespace WorkoutTracker.Core.Implementation.Queries
+namespace WorkoutTracker.Core.Implementation.Queries.Concrete
 {
     public class WorkoutTemplateQuery : IQuery<IEnumerable<WorkoutTemplate>>
     {
