+++ WebApplication/WorkoutTracker.Contracts/QueryDispatcher/Concrete/QueryDispatcher.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using WorkoutTracker.Core.Implementation.Queries.Abstract;
+using WorkoutTracker.Core.Implementation.QueryDispatcher.Abstract;
 
 namespace WorkoutTracker.Core.Implementation.QueryDispatcher.Concrete
 {
